refactor : 2단계 코멘트를 바탕으로 리팩토링 진행 (이너클래스 수정, 접근 제어자 수정)
